requirements.txt
asgiref==3.9.2
Django==5.2.6
sqlparse==0.5.3
psycopg2-binary==2.9.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
argon2-cffi==23.1.0
celery==5.3.4
redis==5.0.1
stripe==7.4.0
requests==2.31.0
django-redis==5.4.0
gunicorn==21.2.0
whitenoise==6.6.0
pillow==10.4.0
django-filter==23.5
drf-spectacular==0.27.0
